+++ docs/changelog.md
@@ -1,5 +1,10 @@
 🔖 [Documentation Home](../README.md) > Changelog
 
+# 1.5.13
+
+- Add note to task execution error for better traceability
+- Implement lazy load for all pydantic ai and openai dependency
+
 # 1.5.12
 
 - Introduce `append_tool` and `append_mcp_server` to `LLMTask`